% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lines.R
\name{clean_lines}
\alias{clean_lines}
\title{Cleans code lines from whitespaces, comments, etc.}
\usage{
clean_lines(all_lines_raw = c(), verbose = FALSE)
}
\arguments{
\item{all_lines_raw}{a character vector containing code lines}

\item{verbose}{a boolean, if TRUE logging will be printed}
}
\value{
a character vector with cleand lines of code
}
\description{
Cleans code lines from whitespaces, comments, etc.
}
\examples{
all_lines_raw <- c("", "# a comment", "x <- 2 + 2", "  print(x)  ")
clean_lines(all_lines_raw)
# [1] "x <- 2 + 2" "print(x)"
}
