% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pkg_info.R
\name{extract_pkg_info}
\alias{extract_pkg_info}
\title{Extract package information}
\usage{
extract_pkg_info(pkg)
}
\arguments{
\item{pkg}{a list with one package information from a renv lockfile.}
}
\value{
a data.frame with the package, its dependencies and the source
}
\description{
A short wrapper function to get a packages dependencies and it's source.
The input is one list element from `renv::lockfile_read()`.
It should contain:

* Package: the package name
* Source: where the packages was sourced, this needs the additional info from the repository entry
* Repository: additional info about the package source
* Requirements: the dependencies of the package
}
\examples{
pkg <- list(Package ="DBI", Source = "Repository", Repository = "CRAN",
            Requirements = c("R", "methods"))
extract_pkg_info(pkg = pkg)

#   pkg     dep source
# 1 DBI       R   CRAN
# 2 DBI methods   CRAN

}
