% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findredundance.R
\name{findredundance}
\alias{findredundance}
\title{Find percentage of redundant code}
\usage{
findredundance(
  dir = NULL,
  script = NULL,
  min_chunk = 2,
  exclude = c("/renv/"),
  verbose = FALSE
)
}
\arguments{
\item{dir}{a folder path, which should be checked}

\item{script}{a character vector with script inputs}

\item{min_chunk}{an integer with the minimum chunk size to be tested}

\item{exclude}{a vector with regex patterns, that should be ignored when reading in `dir`}

\item{verbose}{a boolean setting the debugging prints.}
}
\value{
a data.frame with the summary of repeatable chunks
}
\description{
Find percentage of redundant code
}
