% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_renv_dep.R
\name{plot_renv_dep}
\alias{plot_renv_dep}
\title{Plotting the dependencies of packages in a renv.lock file}
\usage{
plot_renv_dep(lockfile = "renv.lock", dep_level = 2)
}
\arguments{
\item{lockfile}{a file path to a renv.lock file}
}
\value{
a plot
}
\description{
Plotting the dependencies of packages in a renv.lock file
}
